Move make request listener from RequestBuilderExt to RequestListener
